+++ src/Recyclarr.Core/Repo/GitRepositorySource.cs
@@ -8,4 +8,5 @@ public record GitRepositorySource
     public required Uri CloneUrl { get; init; }
     public string Reference { get; init; } = "master";
     public required IDirectoryInfo Path { get; init; }
+    public int Depth { get; init; }
 }