+++ src/Recyclarr.Core/VersionControl/IGitRepositoryFactory.cs
@@ -7,6 +7,8 @@ public interface IGitRepositoryFactory
     Task<IGitRepository> CreateAndCloneIfNeeded(
         Uri repoUrl,
         IDirectoryInfo repoPath,
+        string reference,
+        int depth,
         CancellationToken token
     );
 }